SPI, DAC fixes
DAC initialise fn didn't have chip select brought low, so was using a different reference voltage as a result.
SPI was checking equality against 1 when it should have been comparing against non-zero to determine the value of the MOSI line.
Strange issue with smclk, when it's enabled the MSP seems to go into a loop immediately after it's enabled. Switch to using aclk instead (slower but oh well).
Move to relative percent difference instead of accuracy to avoid issues with division by zero.

+++ src/adc.rs
@@ -106,7 +106,7 @@ impl<CsPin: ADCCSPin, SensorType:ADCSensor> ADC<CsPin, SensorType>{
         result
     }
     pub fn count_to_voltage(&self, count: u16) -> u16{
-        count * self.vcc_millivolts / ADC_RESOLUTION
+        ((count as u32 * self.vcc_millivolts as u32) / ADC_RESOLUTION as u32) as u16
     }
     pub fn read_voltage_from(&mut self, wanted_sensor: &SensorType, spi_bus: &mut impl PayloadSPI<IdleHigh, SampleFirstEdge>) -> u16{
         let count = self.read_count_from(&wanted_sensor, spi_bus);
